add the reassingment of migrants to a daedalus module (#9)
